+++ sdk/resourcemanager/datafactory/armdatafactory/constants.go
@@ -10,7 +10,7 @@ package armdatafactory
 
 const (
 	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/datafactory/armdatafactory"
-	moduleVersion = "v7.0.0"
+	moduleVersion = "v8.0.0"
 )
 
 // ActivityOnInactiveMarkAs - Status result of the activity when the state is set to Inactive. This is an optional property
@@ -1475,22 +1475,6 @@ func PossibleScriptActivityParameterTypeValues() []ScriptActivityParameterType {
 		ScriptActivityParameterTypeSingle,
 		ScriptActivityParameterTypeString,
 		ScriptActivityParameterTypeTimespan,
-	}
-}
-
-// ScriptType - The type of the query. Type: string.
-type ScriptType string
-
-const (
-	ScriptTypeNonQuery ScriptType = "NonQuery"
-	ScriptTypeQuery    ScriptType = "Query"
-)
-
-// PossibleScriptTypeValues returns the possible values for the ScriptType const type.
-func PossibleScriptTypeValues() []ScriptType {
-	return []ScriptType{
-		ScriptTypeNonQuery,
-		ScriptTypeQuery,
 	}
 }